+++ tools/webTools/lib/src/initWebPage.js
@@ -16,8 +16,73 @@ window[MODULE] = [];
 
 // window[MODULE].push();
 // window[MODULE].push();
-// window[MODULE].push();
-// window[MODULE].push();
+/**
+ * 发通知
+ * @type {String}
+ */
+window[MODULE].push(`
+  <div class="modal fade bd-example-modal-lg" id="postNotice" tabindex="-1" role="dialog" aria-labelledby="myLargeModalLabel" aria-hidden="true">
+    <div class="modal-dialog modal-lg">
+      <div class="modal-content">
+        <div class="modal-header">
+            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+              <span class="sr-only">Close</span>
+            </button>
+            <h4 class="modal-title">发送系统通知</h4>
+        </div>
+        <div class="modal-body">
+          <form>
+            <div class="form-group">
+                <label>通知内容</label>
+                <textarea class="form-control" rows="3"></textarea>
+            </div>
+          </form>
+        </div>
+        <div class="modal-footer">
+            <button type="button" class="btn btn-outline-secondary" data-dismiss="modal">关闭</button>
+            <button type="button" class="btn btn-outline-success">发布</button>
+        </div>
+      </div>
+    </div>
+  </div>
+`);
+
+/**
+ * 发公告
+ * @type {String}
+ */
+window[MODULE].push(`
+<div class="modal fade bd-example-modal-lg" id="postAnnouncements" tabindex="-1" role="dialog" aria-labelledby="myLargeModalLabel" aria-hidden="true">
+  <div class="modal-dialog modal-lg">
+    <div class="modal-content">
+      <div class="modal-header">
+          <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+            <span aria-hidden="true">&times;</span>
+            <span class="sr-only">Close</span>
+          </button>
+          <h4 class="modal-title">发布公告</h4>
+      </div>
+      <div class="modal-body">
+        <form>
+          <div class="form-group">
+            <label>公告标题</label>
+            <input type="text" class="form-control" placeholder="请输入公告标题">
+          </div>
+          <div class="form-group">
+              <label>公告正文</label>
+              <textarea class="form-control" rows="10"></textarea>
+          </div>
+        </form>
+      </div>
+      <div class="modal-footer">
+          <button type="button" class="btn btn-outline-secondary" data-dismiss="modal">关闭</button>
+          <button type="button" class="btn btn-outline-success">发布</button>
+      </div>
+    </div>
+  </div>
+</div>
+`);
 
 /**
  * 工作流节点管理